test(embedded/sql): unit testing type inference on value expressions
