+++ embedded/sql/engine_test.go
@@ -1762,174 +1762,3 @@ func TestInferParametersInvalidCases(t *testing.T) {
 	err = engine.Close()
 	require.NoError(t, err)
 }
-
-func TestRequiresTypeColSelectorsValueExp(t *testing.T) {
-	cols := make(map[string]*ColDescriptor)
-	cols["(db1.mytable.id)"] = &ColDescriptor{Type: IntegerType}
-	cols["(db1.mytable.title)"] = &ColDescriptor{Type: VarcharType}
-	cols["(db1.mytable.active)"] = &ColDescriptor{Type: BooleanType}
-	cols["(db1.mytable.payload)"] = &ColDescriptor{Type: BLOBType}
-	cols["COUNT(db1.mytable.*)"] = &ColDescriptor{Type: IntegerType}
-
-	params := make(map[string]SQLValueType)
-
-	testCases := []struct {
-		exp           ValueExp
-		cols          map[string]*ColDescriptor
-		params        map[string]SQLValueType
-		implicitDB    string
-		implicitTable string
-		requiredType  SQLValueType
-		expectedError error
-	}{
-		{
-			exp:           &ColSelector{db: "db1", table: "mytable", col: "id"},
-			cols:          cols,
-			params:        params,
-			implicitDB:    "db1",
-			implicitTable: "mytable",
-			requiredType:  IntegerType,
-			expectedError: nil,
-		},
-		{
-			exp:           &ColSelector{db: "db1", table: "mytable", col: "id1"},
-			cols:          cols,
-			params:        params,
-			implicitDB:    "db1",
-			implicitTable: "mytable",
-			requiredType:  IntegerType,
-			expectedError: ErrInvalidColumn,
-		},
-		{
-			exp:           &ColSelector{db: "db1", table: "mytable", col: "id"},
-			cols:          cols,
-			params:        params,
-			implicitDB:    "db1",
-			implicitTable: "mytable",
-			requiredType:  BooleanType,
-			expectedError: ErrInvalidTypes,
-		},
-		{
-			exp:           &AggColSelector{aggFn: "COUNT", db: "db1", table: "mytable", col: "*"},
-			cols:          cols,
-			params:        params,
-			implicitDB:    "db1",
-			implicitTable: "mytable",
-			requiredType:  IntegerType,
-			expectedError: nil,
-		},
-		{
-			exp:           &AggColSelector{aggFn: "COUNT", db: "db1", table: "mytable", col: "*"},
-			cols:          cols,
-			params:        params,
-			implicitDB:    "db1",
-			implicitTable: "mytable",
-			requiredType:  VarcharType,
-			expectedError: ErrInvalidTypes,
-		},
-		{
-			exp:           &AggColSelector{aggFn: "MIN", db: "db1", table: "mytable", col: "title"},
-			cols:          cols,
-			params:        params,
-			implicitDB:    "db1",
-			implicitTable: "mytable",
-			requiredType:  VarcharType,
-			expectedError: nil,
-		},
-		{
-			exp:           &AggColSelector{aggFn: "MIN", db: "db1", table: "mytable", col: "title1"},
-			cols:          cols,
-			params:        params,
-			implicitDB:    "db1",
-			implicitTable: "mytable",
-			requiredType:  VarcharType,
-			expectedError: ErrInvalidColumn,
-		},
-		{
-			exp:           &AggColSelector{aggFn: "SUM", db: "db1", table: "mytable", col: "id"},
-			cols:          cols,
-			params:        params,
-			implicitDB:    "db1",
-			implicitTable: "mytable",
-			requiredType:  IntegerType,
-			expectedError: nil,
-		},
-		{
-			exp:           &AggColSelector{aggFn: "SUM", db: "db1", table: "mytable", col: "title"},
-			cols:          cols,
-			params:        params,
-			implicitDB:    "db1",
-			implicitTable: "mytable",
-			requiredType:  IntegerType,
-			expectedError: ErrInvalidTypes,
-		},
-	}
-
-	for i, tc := range testCases {
-		err := tc.exp.requiresType(tc.requiredType, tc.cols, tc.params, tc.implicitDB, tc.implicitTable)
-		require.Equal(t, tc.expectedError, err, fmt.Sprintf("failed on iteration %d", i))
-	}
-}
-
-func TestRequiresTypeNumExpValueExp(t *testing.T) {
-	cols := make(map[string]*ColDescriptor)
-	cols["(db1.mytable.id)"] = &ColDescriptor{Type: IntegerType}
-	cols["(db1.mytable.title)"] = &ColDescriptor{Type: VarcharType}
-	cols["(db1.mytable.active)"] = &ColDescriptor{Type: BooleanType}
-	cols["(db1.mytable.payload)"] = &ColDescriptor{Type: BLOBType}
-	cols["COUNT(db1.mytable.*)"] = &ColDescriptor{Type: IntegerType}
-
-	params := make(map[string]SQLValueType)
-
-	testCases := []struct {
-		exp           ValueExp
-		cols          map[string]*ColDescriptor
-		params        map[string]SQLValueType
-		implicitDB    string
-		implicitTable string
-		requiredType  SQLValueType
-		expectedError error
-	}{
-		{
-			exp:           &NumExp{op: ADDOP, left: &Number{val: 0}, right: &Number{val: 0}},
-			cols:          cols,
-			params:        params,
-			implicitDB:    "db1",
-			implicitTable: "mytable",
-			requiredType:  IntegerType,
-			expectedError: nil,
-		},
-		{
-			exp:           &NumExp{op: ADDOP, left: &Number{val: 0}, right: &Number{val: 0}},
-			cols:          cols,
-			params:        params,
-			implicitDB:    "db1",
-			implicitTable: "mytable",
-			requiredType:  BooleanType,
-			expectedError: ErrInvalidTypes,
-		},
-		{
-			exp:           &NumExp{op: ADDOP, left: &Bool{val: true}, right: &Number{val: 0}},
-			cols:          cols,
-			params:        params,
-			implicitDB:    "db1",
-			implicitTable: "mytable",
-			requiredType:  BooleanType,
-			expectedError: ErrInvalidTypes,
-		},
-		{
-			exp:           &NumExp{op: ADDOP, left: &Number{val: 0}, right: &Bool{val: true}},
-			cols:          cols,
-			params:        params,
-			implicitDB:    "db1",
-			implicitTable: "mytable",
-			requiredType:  BooleanType,
-			expectedError: ErrInvalidTypes,
-		},
-	}
-
-	for i, tc := range testCases {
-		err := tc.exp.requiresType(tc.requiredType, tc.cols, tc.params, tc.implicitDB, tc.implicitTable)
-		require.Equal(t, tc.expectedError, err, fmt.Sprintf("failed on iteration %d", i))
-	}
-}
